add processContainer
Problem: Volumes containing other volumes were being processed as
assemblies. This is wrong. Assemblies are physical palacements of solids
that can be moved or rotated or scaled together. Every element in an
Assembly is a sibling of every other element. Containing volumes, on the
oother hand, are volumes that contain other volumes. The containing
volume is a mother of the remaining elements, not a sibling.

Solution: introduced a test in ProcessVolAssem that tests whether the
first element in several objects under a logical volume is a
FeaturePython. If so, a volume is created for that solid and it is
assumed to be a containing volume. The the remaining elements are
process with a new method processContainer, with the first volume as the
parent.

Also, the color for the containing volume is omitted.