merge with Keith test branch of 2023-12-01
- Updated my files to match Keith's test branch
- Minor Python style issues modified from the test branch,
for example solidName = None, changed to solidName=None, as per PEP8
- I had streamlined the create booleans code in GDMLCommands.py, but
those changes were not made by Keith in test, so I inserted those change
into GDMLCommands.py. Because the old code for creating boolenas was
almost identical for each of the boolenas, except for the words "Cut",
"Add", "Intersect", it made sense to derive the code from a comman class
and just adjust that text.